pyproject.toml
[tool.poetry]
name = "job-queue-lambda"
version = "0.0.1.dev0"
description = ""
authors = ["weihong.xu <xuweihong.cn@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "job_queue_lambda"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.10.5"
fire = "^0.7.0"
asyncssh = "^2.19.0"
aiohttp = "^3.11.11"
aiohttp-socks = "0.10.1"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
